Create CoalecentTree example module
